requirements.txt
flask
flask_cors
pymongo
pyjwt